Update 10824_FourNumbers.cpp
마지막 문자열 결과를 숫자로 변환해 주는 부분을 int -> long long 으로 변경

Result : PASS
Memory : 2028 KB
Time : 0 ms
Code size : 873 B

+++ BJCodingTest/AlgorithmBasic/10824_FourNumbers.cpp
@@ -40,7 +40,7 @@ int main()
     auto num0 = splitStr[0] + splitStr[1];
     auto num1 = splitStr[2] + splitStr[3];
 
-    auto sum = std::stoi(num0) + std::stoi(num1);
+    auto sum = std::stoll(num0) + std::stoll(num1);
     std::cout << sum << "\n";
 
     return 0;
